uncap fastapi dependency
`fastapi` is a project with a very frequent release cadence, and there's little risk of their upstream improvements breaking things for this project. Users of this excellent tool should be able to update their `fastapi` install without having to also update this library, especially when there are significant breaking changes like the recent 9x-10x improvements.